pyproject.toml
[tool.poetry]
name = "morninghour"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.7"
postgres = "^4.0"
djangorestframework = "^3.14.0"
Markdown = "^3.5.1"
django-filter = "^23.4"
Pillow = "^10.1.0"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"
django-leaflet = "^0.29.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
